Refactor code for better organization
Moved "Expr" and "Insn" enums from cfg.rs to newly created expr.rs and insn.rs files respectively. This will make the components more manageable and improve readability of the code. Also, added 'mod expr' and 'mod insn' in mod.rs to include newly created files in the module.
