+++ src/ssa_ir/mod.rs
@@ -1,1 +1,3 @@
 mod cfg;
+mod expr;
+mod insn;
